requirements.txt
selenium
pytest
allure-pytest
PyHamcrest
addict